+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+### Version 0.0.2
+
+- support for OSX
+- use the right path for setting up the workspace
+- fixed wrong label decomposition
+- only consider c++ targets for c_cpp_properties.json generation
+
 ### Version 0.0.1
 
 - Preview release (experimental)